Implement data layer with in memory H2 database

+++ src/main/java/com/nashvi/labs/accounts/AccountNotFoundException.java
@@ -8,7 +8,7 @@ public class AccountNotFoundException extends RuntimeException {
 
 	private static final long serialVersionUID = 1L;
 
-	public AccountNotFoundException(int id) {
+	public AccountNotFoundException(Long id) {
 		super("could not find account '" + id + "'.");
 	}
